requirements.txt
# Core deep learning frameworks
torch>=1.9.0
torchvision>=0.10.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0

# Scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0

# Progress bars
tqdm>=4.62.0

# Perceptual loss (optional)
lpips>=0.1.4

# Data handling
glob2>=0.7